Move DragAndDrop component and style to match current structure
