+++ app/routes/DragAndDrop/DragAndDrop.js
@@ -1,17 +1,14 @@
-
-
 import React, {Component} from "react";
 
 import {
-    StyleSheet,
     View,
     Text,
     PanResponder,
     Animated,
-    Easing,
-    Dimensions
+    Easing
 } from 'react-native';
 
+import style from './style.js'
 
 export default class DragAndDrop extends Component{
     constructor(props){
@@ -57,11 +54,11 @@ export default class DragAndDrop extends Component{
 
     render(){
         return (
-            <View style={styles.mainContainer}>
+            <View style={style.mainContainer}>
                 <View 
                     onLayout={this.setDropZoneValues.bind(this)}
-                    style={styles.dropZone}>
-                    <Text style={styles.text}>Drop me here!</Text>
+                    style={style.dropZone}>
+                    <Text style={style.text}>Drop me here!</Text>
                 </View>
 
                 {this.renderDraggable()}
@@ -72,44 +69,14 @@ export default class DragAndDrop extends Component{
     renderDraggable(){
         if(this.state.showDraggable){
             return (
-                <View style={styles.draggableContainer}>
+                <View style={style.draggableContainer}>
                     <Animated.View 
                         {...this.panResponder.panHandlers}
-                        style={[this.state.pan.getLayout(), styles.circle]}>
-                        <Text style={styles.text}>Drag me!</Text>
+                        style={[this.state.pan.getLayout(), style.circle]}>
+                        <Text style={style.text}>Drag me!</Text>
                     </Animated.View>
                 </View>
             );
         }
     }
 }
-
-let CIRCLE_RADIUS = 36;
-let Window = Dimensions.get('window');
-let styles = StyleSheet.create({
-    mainContainer: {
-        flex    : 1
-    },
-    dropZone    : {
-        height  : 100,
-        backgroundColor:'#2c3e50'
-    },
-    text        : {
-        marginTop   : 25,
-        marginLeft  : 5,
-        marginRight : 5,
-        textAlign   : 'center',
-        color       : '#fff'
-    },
-    draggableContainer: {
-        position    : 'absolute',
-        top         : Window.height/2 - CIRCLE_RADIUS,
-        left        : Window.width/2 - CIRCLE_RADIUS,
-    },
-    circle      : {
-        backgroundColor     : '#1abc9c',
-        width               : CIRCLE_RADIUS*2,
-        height              : CIRCLE_RADIUS*2,
-        borderRadius        : CIRCLE_RADIUS
-    }
-});
